+++ packages/react-next/src/components/SwatchColorPicker/ColorPickerGridCell.base.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { ITheme, mergeStyleSets, IProcessedStyleSet } from '../../Styling';
 import { classNamesFunction, memoizeFunction } from '../../Utilities';
 import { getColorFromString } from 'office-ui-fabric-react/lib/Color';
-import { GridCell } from '../../utilities/grid/GridCell';
+import { ButtonGridCell } from '../../utilities/ButtonGrid/ButtonGridCell';
 import { getStyles as getActionButtonStyles } from 'office-ui-fabric-react/lib/components/Button/ActionButton/ActionButton.styles';
 import { IButtonClassNames } from 'office-ui-fabric-react/lib/components/Button/BaseButton.classNames';
 import {
@@ -110,7 +110,7 @@ export const ColorPickerGridCellBase: React.FunctionComponent<IColorPickerGridCe
   };
 
   return (
-    <GridCell
+    <ButtonGridCell
       item={item}
       id={`${idPrefix}-${item.id}-${item.index}`}
       key={item.id}
