+++ packages/react-next/src/utilities/ButtonGrid/ButtonGrid.styles.ts
@@ -1,0 +1,13 @@
+import { IButtonGridStyleProps, IButtonGridStyles } from './ButtonGrid.types';
+
+export const getStyles = (props: IButtonGridStyleProps): IButtonGridStyles => {
+  return {
+    root: {
+      padding: 2,
+      outline: 'none',
+    },
+    tableCell: {
+      padding: 0,
+    },
+  };
+};
