--- packages/react-next/src/utilities/grid/Grid.tsx
@@ -1,6 +1,0 @@
-import { styled } from '../../Utilities';
-import { GridBase } from './Grid.base';
-import { IGridProps, IGridStyleProps, IGridStyles } from './Grid.types';
-import { getStyles } from './Grid.styles';
-
-export const Grid = styled<IGridProps, IGridStyleProps, IGridStyles, HTMLElement>(GridBase, getStyles);
