--- packages/react-next/src/utilities/grid/index.ts
@@ -1,4 +1,0 @@
-export * from './Grid';
-export * from './Grid.types';
-export * from './GridCell';
-export * from './GridCell.types';
